Add files via upload
Geogebra files for constructing P-Ceva-Conjugate of U by computation of trilinear coordinates. And then showing P-Ceva-Conjugate of U is not same as U-Ceva-Conjugate of P, unless P=U. Also P-Ceva-Conjugate of P-Ceva-Conjugate of U is U. That is, the operator P-Ceva-Conjugate is a transformation that is its own inverse. Applying the transform twice gives back the original argument U. So for a given P, the transform P-Ceva-Conjugate is an involution.