Bianticevian: When 4 foci are concyclic.
Locus of Q, when the foci F,G of the 2 parabolas through a given P circumscribing its anticevian Triangle, and the foci (D,E) of the bianticevian conic C(P,Q) are concyclic:
The locus of Q is two different hyperbolas shown in orange. The corresponding 2 different circles are also shown in orange.